requirements.txt
Django
djangorestframework
psycopg2-binary
djangorestframework-simplejwt
django_extensions